requirements.txt
fastapi==0.112.1
openai==1.42.0
python-dotenv==1.0.1
tiktoken==0.7.0
